+++ src/test/scala/org/broadinstitute/hail/methods/GQByDPBinSuite.scala
@@ -1,6 +1,7 @@
 package org.broadinstitute.hail.methods
 
 import org.broadinstitute.hail.SparkSuite
+import org.broadinstitute.hail.io.vcf.LoadVCF
 import org.testng.annotations.Test
 
 class GQByDPBinSuite extends SparkSuite {