+++ src/test/scala/org/broadinstitute/hail/methods/PedigreeSuite.scala
@@ -2,6 +2,7 @@ package org.broadinstitute.hail.methods
 
 import org.broadinstitute.hail.SparkSuite
 import org.broadinstitute.hail.check.Prop._
+import org.broadinstitute.hail.io.vcf.LoadVCF
 import org.testng.annotations.Test